+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### New
 - Metrics modules (Dagger, Data, Domain).
 
+### Fixes
+
+- Fixed Login for Product VPN with TwoPassMode enabled.
+
 ## [7.1.0]
 
 ### New